Add files via upload
Description:
This social media influencers analysis project provides a detailed exploration of the behavior and effectiveness of various influencers on platforms such as Instagram, YouTube, and TikTok. Through comprehensive data analysis, visualizations, and statistical modeling, the aim is to understand the impact of influencers on the audience and their ability to drive conversions and engagement.

Types of Analysis:

Correlation analysis between key metrics such as followers, engagement, and conversion rates.
Segmentation and clustering of influencers based on performance metrics.
Survival analysis to identify the follower loss rate over time.
Conversion rate analysis to assess the effectiveness of influencers in generating conversions.
Data visualization to present findings and patterns effectively.
Requisites to visualize the information:

RStudio: An IDE for R that facilitates data exploration, visualization creation, and statistical analysis execution.
R Packages: Install the required packages mentioned in the project to be able to execute the analyses and visualizations.
To get started, ensure you have RStudio installed and set up this project (.Rproj) in your workspace. Next, install the required packages using the install.packages() function and run the provided analysis scripts to gain insights into the social media influencers' performance.

Explore the data and uncover key trends driving the success of influencers on social media!